VR: split files
